Fix MPLY-7721 'Floor picker UI is absent for multi-story buildings.' For interiors explorer view, pixel scale was being incorrectly cast to int and truncated. For iphone 6+, pixel scale is non-integer (2.60869575), so resulted in wrong layout Buddy: Michael

+++ ios/ios_src/InteriorsExplorer/View/InteriorsExplorerViewModule.mm
@@ -18,9 +18,9 @@ namespace ExampleApp
                                                                      const Eegeo::Rendering::ScreenProperties& screenProperties,
                                                                      Eegeo::Helpers::IdentityProvider& identityProvider)
             {
-                const int screenWidth = screenProperties.GetScreenWidth();
-                const int screenHeight = screenProperties.GetScreenHeight();
-                const int pixelScale = screenProperties.GetPixelScale();
+                const float screenWidth = screenProperties.GetScreenWidth();
+                const float screenHeight = screenProperties.GetScreenHeight();
+                const float pixelScale = screenProperties.GetPixelScale();
                 
                 m_pTutorialView = [[InteriorsExplorerTutorialView alloc] initWithParams: screenWidth : screenHeight : pixelScale];
